+++ lib/workerHelper.ts
@@ -1,19 +1,19 @@
-'use babel';
-
-// eslint-disable-next-line import/extensions, import/no-extraneous-dependencies
-import { Task } from 'atom';
+import { Task, TextEditor } from 'atom';
+import type { ConfigSchema } from "./config"
 import cryptoRandomString from 'crypto-random-string';
+import type * as Tslint from "tslint";
 
-export default class WorkerHelper {
+export class WorkerHelper {
+  workerInstance: Task | null
   constructor() {
     this.workerInstance = null;
   }
 
   isRunning() {
-    return !!this.workerInstance;
+    return Boolean(this.workerInstance);
   }
 
-  startWorker(config) {
+  startWorker(config: ConfigSchema) {
     if (!this.workerInstance) {
       this.workerInstance = new Task(require.resolve('./worker.js'));
       this.workerInstance.start(config);
@@ -27,23 +27,26 @@ export default class WorkerHelper {
     }
   }
 
-  changeConfig(key, value) {
+  changeConfig(key: string, value: any) {
     if (this.workerInstance) {
       this.workerInstance.send({
         messageType: 'config',
         message: { key, value },
-      });
+      } as ConfigMessage);
     }
   }
 
-  requestJob(jobType, textEditor) {
-    if (!this.workerInstance) {
+  async requestJob(jobType: string, textEditor: TextEditor): Promise<Tslint.LintResult[]> {
+    if (this.workerInstance === null) {
       throw new Error("Worker hasn't started");
     }
 
-    const emitKey = cryptoRandomString({ length: 10 });
+    const emitKey = await cryptoRandomString.async({ length: 10 });
 
     return new Promise((resolve, reject) => {
+      if (this.workerInstance === null) {
+        throw new Error("Worker hasn't started");
+      }
       const errSub = this.workerInstance.on('task:error', (...err) => {
         // Re-throw errors from the task
         const error = new Error(err[0]);
@@ -52,7 +55,7 @@ export default class WorkerHelper {
         reject(error);
       });
 
-      const responseSub = this.workerInstance.on(emitKey, (data) => {
+      const responseSub = this.workerInstance.on(emitKey, (data: Tslint.LintResult[]) => {
         errSub.dispose();
         responseSub.dispose();
         resolve(data);
@@ -67,10 +70,28 @@ export default class WorkerHelper {
             content: textEditor.getText(),
             filePath: textEditor.getPath(),
           },
-        });
+        } as JobMessage);
       } catch (e) {
         reject(e);
       }
     });
+  }
+}
+
+export type ConfigMessage = {
+  messageType: 'config',
+  message: {
+    key: keyof ConfigSchema,
+    value: boolean | string | null,
+  }
+}
+
+export type JobMessage = {
+  messageType: 'job',
+  message: {
+    emitKey: string,
+    jobType: string,
+    content: ReturnType<TextEditor["getText"]>,
+    filePath: ReturnType<TextEditor["getPath"]>,
   }
 }